+++ aws/curator.yaml
@@ -70,11 +70,6 @@ spec:
                 secretKeyRef:
                   name: curator-dev-bch28c9gkm
                   key: aws_secret_access_key
-            - name: MAPBOX_TOKEN
-              valueFrom:
-                secretKeyRef:
-                  name: curator-dev-bch28c9gkm
-                  key: mapbox_token
             - name: EMAIL_USER_ADDRESS
               valueFrom:
                 secretKeyRef:
@@ -160,11 +155,6 @@ spec:
                 secretKeyRef:
                   name: curator-prod-4mtk8mh9t2
                   key: aws_secret_access_key
-            - name: MAPBOX_TOKEN
-              valueFrom:
-                secretKeyRef:
-                  name: curator-prod-4mtk8mh9t2
-                  key: mapbox_token
             - name: EMAIL_USER_ADDRESS
               valueFrom:
                 secretKeyRef:
