+++ data-serving/data-service/package.json
@@ -54,7 +54,10 @@
     "typescript": "^3.9.7"
   },
   "dependencies": {
+    "@mapbox/mapbox-sdk": "^0.11.0",
     "@types/js-yaml": "^3.12.5",
+    "@types/lru-cache": "^5.1.0",
+    "@types/mapbox__mapbox-sdk": "^0.6.2",
     "@types/multer": "^1.4.4",
     "@types/swagger-ui-express": "^4.1.2",
     "@types/yamljs": "^0.2.31",
@@ -69,6 +72,7 @@
     "express-validator": "^6.6.1",
     "lodash": "^4.17.20",
     "longjohn": "^0.2.12",
+    "lru-cache": "^6.0.0",
     "mongodb": "^3.6.0",
     "mongoose": "^5.10.0",
     "search-query-parser": "^1.5.5",
