+++ data-serving/data-service/schemas/cases.indexes.json
@@ -184,5 +184,15 @@
             "locale": "en_US",
             "strength": 2
         }
+    },
+    {
+        "name": "locationQueryIdx",
+        "key": {
+            "location.query": -1
+        },
+        "collation": {
+            "locale": "en_US",
+            "strength": 2
+        }
     }
 ]
