+++ data-serving/data-service/src/util/validate-env.ts
@@ -1,9 +1,12 @@
-import { CleanEnv, cleanEnv, port, str } from 'envalid';
+import { CleanEnv, cleanEnv, port, str, bool } from 'envalid';
 
 export default function validateEnv(): Readonly<{
     DB_CONNECTION_STRING: string;
     PORT: number;
     SERVICE_ENV: string;
+    MAPBOX_PERMANENT_GEOCODE: boolean;
+    MAPBOX_TOKEN: string;
+    ENABLE_FAKE_GEOCODER: boolean;
 }> &
     CleanEnv & {
         readonly [varName: string]: string | undefined;
@@ -19,6 +22,20 @@ export default function validateEnv(): Readonly<{
             choices: ['local', 'dev', 'prod'],
             desc: 'Environment in which the service is running',
             devDefault: 'local',
+        }),
+        MAPBOX_PERMANENT_GEOCODE: bool({
+            desc: 'Whether to use the permanent geocode endpoint',
+            devDefault: false,
+            default: true,
+        }),
+        MAPBOX_TOKEN: str({
+            desc: 'Mapbox token to use for geocoding',
+            devDefault: '',
+        }),
+        ENABLE_FAKE_GEOCODER: bool({
+            desc: 'Whether to enable the fake seedable geocoder',
+            devDefault: true,
+            default: false,
         }),
     });
 }
