+++ data-serving/data-service/test/geocoding/suggest.test.ts
@@ -1,9 +1,6 @@
-import * as baseUser from '../users/base.json';
-
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import app from '../../src/index';
 import request from 'supertest';
-import supertest from 'supertest';
 
 let mongoServer: MongoMemoryServer;
 
@@ -23,13 +20,8 @@ describe('Geocodes', () => {
             geometry: { latitude: 45.75889, longitude: 4.84139 },
             name: 'Lyon',
         };
-        request(app).post('/api/geocode/seed').send(lyon).expect(200);
-        const curatorRequest = supertest.agent(app);
-        await curatorRequest
-            .post('/auth/register')
-            .send({ ...baseUser, ...{ roles: ['curator'] } })
-            .expect(200);
-        curatorRequest
+        await request(app).post('/api/geocode/seed').send(lyon).expect(200);
+        return request(app)
             .get('/api/geocode/suggest')
             .query({ q: 'Lyon' })
             .expect(200, [lyon]);
