+++ data-serving/data-service/test/model/data/location.full.json
@@ -9,5 +9,6 @@
     "geometry": {
         "latitude": 40.6809611,
         "longitude": -73.9740028
-    }
+    },
+    "query": "Brooklyn"
 }
