+++ verification/curator-service/api/package.json
@@ -57,9 +57,6 @@
     "typescript": "^3.9.7"
   },
   "dependencies": {
-    "@mapbox/mapbox-sdk": "github:mapbox/mapbox-sdk-js",
-    "@types/lru-cache": "^5.1.0",
-    "@types/mapbox__mapbox-sdk": "^0.6.2",
     "@types/multer": "^1.4.4",
     "@types/swagger-ui-express": "^4.1.2",
     "@types/yamljs": "^0.2.31",
@@ -74,7 +71,6 @@
     "express-session": "^1.17.1",
     "express-status-monitor": "^1.3.3",
     "longjohn": "^0.2.12",
-    "lru-cache": "^6.0.0",
     "mongoose": "^5.10.0",
     "nodemailer": "^6.4.11",
     "passport": "^0.4.1",
